renamed controller, moved validated structs into controller/mod.rs, added validated cluster usage in place of raw cluster in some places
