+++ rust/operator-binary/src/service.rs
@@ -8,7 +8,10 @@ use stackable_operator::{
     role_utils::RoleGroupRef,
 };
 
-use crate::crd::{HTTPS_PORT, HTTPS_PORT_NAME, METRICS_PORT, METRICS_PORT_NAME, v1alpha1};
+use crate::{
+    controller::ValidatedCluster,
+    crd::{HTTPS_PORT, HTTPS_PORT_NAME, METRICS_PORT, METRICS_PORT_NAME, v1alpha1},
+};
 
 #[derive(Snafu, Debug)]
 pub enum Error {
@@ -26,18 +29,18 @@ pub enum Error {
 /// The rolegroup headless [`Service`] is a service that allows direct access to the instances of a certain rolegroup
 /// This is mostly useful for internal communication between peers, or for clients that perform client-side load balancing.
 pub fn build_rolegroup_headless_service(
-    nifi: &v1alpha1::NifiCluster,
+    cluster: &ValidatedCluster,
     role_group_ref: &RoleGroupRef<v1alpha1::NifiCluster>,
-    object_labels: ObjectLabels<v1alpha1::NifiCluster>,
+    object_labels: &ObjectLabels<ValidatedCluster>,
     selector: BTreeMap<String, String>,
 ) -> Result<Service, Error> {
     Ok(Service {
         metadata: ObjectMetaBuilder::new()
-            .name_and_namespace(nifi)
+            .name_and_namespace(cluster)
             .name(role_group_ref.rolegroup_headless_service_name())
-            .ownerreference_from_resource(nifi, None, Some(true))
+            .ownerreference_from_resource(cluster, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(&object_labels)
+            .with_recommended_labels(object_labels)
             .context(MetadataBuildSnafu)?
             .build(),
         spec: Some(ServiceSpec {
@@ -55,19 +58,19 @@ pub fn build_rolegroup_headless_service(
 
 /// The rolegroup metrics [`Service`] is a service that exposes metrics and a prometheus scraping label.
 pub fn build_rolegroup_metrics_service(
-    nifi: &v1alpha1::NifiCluster,
+    cluster: &ValidatedCluster,
     role_group_ref: &RoleGroupRef<v1alpha1::NifiCluster>,
-    object_labels: ObjectLabels<v1alpha1::NifiCluster>,
+    object_labels: &ObjectLabels<ValidatedCluster>,
     selector: BTreeMap<String, String>,
     product_version: &str,
 ) -> Result<Service, Error> {
     Ok(Service {
         metadata: ObjectMetaBuilder::new()
-            .name_and_namespace(nifi)
+            .name_and_namespace(cluster)
             .name(role_group_ref.rolegroup_metrics_service_name())
-            .ownerreference_from_resource(nifi, None, Some(true))
+            .ownerreference_from_resource(cluster, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
-            .with_recommended_labels(&object_labels)
+            .with_recommended_labels(object_labels)
             .context(MetadataBuildSnafu)?
             .with_labels(prometheus_labels())
             .with_annotations(prometheus_annotations(product_version))
